REF: move safe_sort to algos to avoid private/circular dependencies
